requirements.txt

Flask==3.0.3
python-dotenv==1.0.1
python-telegram-bot==13.0
requests==2.31.0
urllib3==1.26.16
instaloader==4.11
google-generativeai
pillow==10.3.0